+++ lib/server.dart
@@ -1,0 +1,21 @@
+library softhai.bulls_eye.Server;
+
+import 'dart:io';
+import 'dart:async';
+import 'common.dart' as common;
+
+// Globals
+part 'src/Server/req_res_context.dart';
+part 'src/Server/Server.dart';
+part 'src/Server/Exceptions.dart';
+
+// Middleware
+part 'src/Server/Middleware/middleware.dart';
+part 'src/Server/Middleware/middleware_channel_part.dart';
+part 'src/Server/Middleware/middleware_error.dart';
+
+// Routing
+part 'src/Server/Routing/RouteManager.dart';
+part 'src/Server/Routing/ExecuteCode.dart';
+part 'src/Server/Routing/LoadFile.dart';
+part 'src/Server/Routing/RoutingExceptions.dart';